+++ CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `smallestImage` in `TidalAlbumResource`
+- `largestImage` in `TidalAlbumResource`
+- `smallestImageWithSizeAtLeast(width:height:)` in `TidalAlbumResource`
+
+### Changed
+
+- `TidalAlbumResource`: `images` renamed to `imageArtwork`
+- `TidalAlbumResource`: `videos` renamed to `videoArtwork`
+
+### Deprecated
+TBD
+
+### Removed
+TBD
+
+### Fixed
+TBD
+
+### Security
+TBD
+
 ## [0.3.0] - 2026-03-10
 
 ### Added